Update Random to gradle
