--- Random/lombok.config
@@ -1,1 +1,0 @@
-lombok.log.custom.declaration = com.ikalagaming.logging.Logger com.ikalagaming.logging.LoggerFactory.getLogger(TOPIC)
