+++ Random/settings.gradle
@@ -1,0 +1,3 @@
+rootProject.name = 'Random'
+include ':Ikala-Core'
+project(':Ikala-Core').projectDir = new File('../../Ikala-Core')
