--- Random/src/com/ikalagaming/random/RandomPlugin.java
@@ -1,11 +1,0 @@
-package com.ikalagaming.random;
-
-import com.ikalagaming.plugins.Plugin;
-
-/**
- * An example plugin to demonstrate how plugins are set up.
- * 
- * @author Ches Burks
- *
- */
-public class RandomPlugin extends Plugin {}
