+++ Random/src/main/resources/plugin.yml
@@ -1,0 +1,5 @@
+name: Random
+version: 0.0.1
+description: Random generation utilities
+authors: [maceris]
+main-class: com.ikalagaming.random.RandomPlugin
